moved fancy edit component to root component directory
