+++ zapread.com/Scripts/src/pages/post/edit.js
@@ -8,7 +8,7 @@ import React, { useCallback, useEffect, useState, useRef } from 'react';// [✓]
 import { Container, Row, Col } from 'react-bootstrap';                  // [✓]
 import ReactDOM from 'react-dom';                                       // [✓]
 import Swal from 'sweetalert2';                                         // [✓]
-import Input from './Components/Input';                                 // [✓]
+import Input from '../../Components/Input/Input';                       // [✓]
 import Editor from './Components/Editor';                               // [✓]
 import DraftsTable from './Components/DraftsTable';                     // [✓]
 import Picker from './Components/Picker';                               // [✓]
